Remove the need for a hash file, separate test and reference creation
