+++ core/util/createBitmaps.js
@@ -1,0 +1,38 @@
+var path = require('path');
+
+var fs = require('./fs');
+var runCasper = require('./runCasper');
+
+function writeReferenceCreateConfig(config, isReference) {
+  var configJSON = require(config.backstopConfigFileName);
+
+  configJSON.isReference = isReference;
+  configJSON.paths.tempCompareConfigFileName = config.tempCompareConfigFileName;
+
+  return fs.writeFile(config.captureConfigFileName, JSON.stringify(configJSON));
+}
+
+module.exports = function(config, isReference) {
+  return writeReferenceCreateConfig(config, isReference).then(function () {
+    var tests = [path.join(config.backstop, 'capture/genBitmaps.js')];
+    var casperChild = runCasper(config, tests);
+
+    return new Promise(function (resolve, reject) {
+      casperChild.on('close', function (code) {
+        var success = code === 0; // Will be 1 in the event of failure
+        var result = (success) ? 'Bitmap file generation completed.' : 'Testing script failed with code: ' + code;
+
+        console.log('\n' + result);
+
+        // exit if there was some kind of failure in the casperChild process
+        if (code !== 0) {
+          console.log('\nLooks like an error occured. You may want to try running `$ backstop echo`. This will echo the requested test URL output to the console. You can check this output to verify that the file requested is indeed being received in the expected format.');
+          reject(new Error('An error occured. You may want to try running `$ backstop echo`.'));
+          return;
+        }
+
+        resolve();
+      });
+    });
+  });
+};
