+++ package.json
@@ -10,7 +10,6 @@
     "genConfig": "node ./cli/index.js genConfig",
     "reference": "node ./cli/index.js reference",
     "test": "node ./cli/index.js test",
-    "bless": "node ./cli/index.js bless",
     "openReport": "node ./cli/index.js openReport",
     "echo": "node ./cli/index.js echo"
   },
@@ -30,7 +29,6 @@
   "dependencies": {
     "casperjs": "^1.1.0-beta5",
     "chalk": "^1.1.3",
-    "checksum": "^0.1.1",
     "copy": "^0.3.0",
     "del": "^2.2.2",
     "es6-promise": "^3.2.1",
